pyproject.toml
[tool.poetry]
name = "drf"
version = "0.1.0"
description = ""
authors = ["Norilka <NorilkaRus@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.1"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.14.0"
pillow = "^10.2.0"
requests = "^2.31.0"
ipython = "^8.20.0"
pytils = "^0.4.1"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
